Revert refactor
